stage 3

+++ src/main/java/view/Main.java
@@ -8,10 +8,12 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
+import static controller.ProjectRegister.deleteParticipant;
 import static controller.UserRegister.getUser;
 
 
 public class Main {
+    //todo пофиксить абсолютные пути
     private static BufferedReader in = new BufferedReader(new InputStreamReader(System.in));
     private static Person user;
     public static UserRegister userRegister = new UserRegister();
@@ -39,24 +41,54 @@ public class Main {
         }
         switch (i) {
             case 0:
-                projectRegister.printProjects();
+                writeCommands();
                 break;
             case 1:
-                userRegister.printUsers();
+                projectRegister.printProjects();
                 break;
             case 2:
+                userRegister.printUsers();
+                break;
+            case 3:
                 System.out.println("Введите тэг для поиска проектов");
                 String tag = in.readLine().trim();
                 projectRegister.search(tag);
                 break;
-            case 3:
+            case 4:
+                UserRegister.sortByRating();
+                break;
+            case 5:
                 System.out.println("Введите подстроку для поиска пользователя");
                 String fragment = in.readLine().trim().toLowerCase();
                 UserRegister.search(fragment);
                 break;
-            case 4:
+            case 6:
+                if (user.permission != Person.Permission.PARTICIPANT) {
+                    input = in.readLine().trim();
+                    projectRegister.addParticipant(input);
+                } else {
+                    System.out.println("У вас нет доступа");
+                }
+                break;
+            case 7:
+                if (user.permission != Person.Permission.PARTICIPANT) {
+                    input = in.readLine().trim();
+                    deleteParticipant(input);
+                } else {
+                    System.out.println("У вас нет доступа");
+                }
+                break;
+            case 8:
+                if (user.permission != Person.Permission.PARTICIPANT) {
+                    input = in.readLine().trim();
+                    UserRegister.changeRating(input);
+                } else {
+                    System.out.println("У вас нет доступа");
+                }
+                break;
+            case 9:
                 try {
-                    if (user.permission == Person.Permission.PROFESSOR || user.permission == Person.Permission.ADMIN) {
+                    if (user.permission == Person.Permission.ADMIN) {
                         System.out.println("Введите данные пользователя");
                         input = in.readLine().trim();
                         userRegister.addPerson(input);
@@ -67,9 +99,16 @@ public class Main {
                     System.out.println("Oops");
                 }
                 break;
-            case 5:
+            case 10:
+                if (user.permission == Person.Permission.ADMIN) {
+                    System.out.println("Введите имя студента, которого надо удалить");
+                    input = in.readLine().trim();
+                    userRegister.deletePerson(input);
+                }
+                break;
+            case 11:
                 try {
-                    if (user.permission == Person.Permission.PROFESSOR || user.permission == Person.Permission.ADMIN) {
+                    if (user.permission == Person.Permission.ADMIN) {
                         System.out.println("Введите данные о проекте");
                         input = in.readLine().trim();
                         projectRegister.addProject(input);
@@ -81,34 +120,45 @@ public class Main {
                     System.out.println("Oops");
                 }
                 break;
-            case 6:
-                input = in.readLine().trim();
-                projectRegister.deleteProject(input);
+            case 12:
+                if (user.permission == Person.Permission.ADMIN) {
+                    input = in.readLine().trim();
+                    projectRegister.deleteProject(input);
+                } else {
+                    System.out.println("У вас нет доступа");
+                }
                 break;
-            case 7:
-                input = in.readLine().trim();
-                projectRegister.addParticipant(input);
-                break;
-            case 8:
-                input = in.readLine().trim();
-                projectRegister.deleteParticipant(input);
-                break;
-            case 9:
+            case 13:
                 authorization();
                 break;
-            case 10:
+            case 14:
                 return;
             default:
                 System.out.println("Такой команды не существует");
-                workWithProjects();
                 break;
         }
         workWithProjects();
-        //todo сортировка проектов, людей
-        //todo добавление проекта, человека
-        //todo оценивание
     }
 
+    private static void writeCommands(){
+        StringBuilder sb = new StringBuilder();
+        sb.append("0 - вывод всех команд;").append("\n");
+        sb.append("1 - вывод всех проектов (доступно всем);").append("\n");
+        sb.append("2 - вывод всех пользователей (доступно всем);").append("\n");
+        sb.append("3 - поиск проекта по тэгам (доступно всем);").append("\n");
+        sb.append("4 - сортировка участников по рейтингу (доступно всем);").append("\n");
+        sb.append("5 - поиск участника по тэгу (доступно всем);").append("\n");
+        sb.append("6 - добавление участника в проект (доступно администратору, профессору);").append("\n");
+        sb.append("7 - удаление участника из проекта (доступно администратору, профессору);").append("\n");
+        sb.append("8 - изменение рейтинга участника (доступно администратору, профессору);").append("\n");
+        sb.append("9 - добавление пользователя в систему (доступно администратору);").append("\n");
+        sb.append("10 - удаление пользователя из системы (доступно администратору);").append("\n");
+        sb.append("11 - добавления проекта в системы (доступно администратору);").append("\n");
+        sb.append("12 - удаление проекта из системы (доступно администратору);").append("\n");
+        sb.append("13 - сменить пользователя;").append("\n");
+        sb.append("14 - завершение работы;").append("\n");
+        System.out.println(sb);
+    }
     public static void main(String[] args) {
         System.out.println("");
         /**
@@ -123,6 +173,7 @@ public class Main {
         }
         //работа с проектами
         try {
+            writeCommands();
             workWithProjects();
         } catch (IOException e) {
             e.printStackTrace();
